refactorings are functional
